Aggregate travel expenses

Aggregate the sf (geometry type point) attributes "Min_Dist" and "Min_Time" to an area of interest

aggregate_expenses(aoi_sf, pnt_sf, id_col, crs = 4647)

Arguments

aoi_sf
sf of geometry type polygon used for aggregation

pnt_sf
sf of geometry type point including attribute to be aggregated

id_col
name of the id column used for grouping operations

crs
epsg code

Value

sf of geometry type polygon including aggregated distance and time values

See also

Examples

data(testdata)
pnt_sf <- create_time_distance_sf(odm_object = testdata$od_result)

#> New names:
#> * `` -> ...1
#> * `` -> ...2

#> New names:
#> * `` -> ...1
#> * `` -> ...2

output_sf <- aggregate_expenses(
  aoi_sf = testdata$grid_500m,
  pnt_sf = pnt_sf,
  id_col = "id",
  crs = 4647
)

#> Joining, by = "id"